+++ router_test.go
@@ -21,11 +21,12 @@
 package yarpc_test
 
 import (
+	"context"
 	"testing"
 
-	"golang.org/x/net/context"
-
 	"go.uber.org/yarpc"
+	"go.uber.org/yarpc/api/middleware"
+	"go.uber.org/yarpc/api/middleware/middlewaretest"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/api/transport/transporttest"
 
@@ -143,4 +144,23 @@ func TestEmptyProcedureRegistration(t *testing.T) {
 	assert.Panics(t,
 		func() { m.Register(procedures) },
 		"expected router panic")
+}
+
+func TestRouterWithMiddleware(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	ctx := context.Background()
+	req := &transport.Request{}
+	expectedSpec := transport.HandlerSpec{}
+
+	routerMiddleware := middlewaretest.NewMockRouter(mockCtrl)
+	routerMiddleware.EXPECT().Choose(ctx, req, gomock.Any()).Times(1).Return(expectedSpec, nil)
+
+	router := middleware.ApplyRouteTable(yarpc.NewMapRouter("service"), routerMiddleware)
+
+	actualSpec, err := router.Choose(ctx, req)
+
+	assert.Equal(t, expectedSpec, actualSpec, "handler spec returned from route table did not match")
+	assert.Nil(t, err)
 }